requirements.txt
torch
torchvision
opencv-python
numpy
pillow
scikit-learn
matplotlib
seaborn
tqdm
pandas